summarizer/serializers.py
from rest_framework import serializers
from summarizer.models import Analysis

class AnalysisSerializer(serializers.ModelSerializer):
    class Meta:
        model = Analysis
        fields = '__all__'
