requirements.txt
django
djangorestframework
openai
python-dotenv
spacy==3.8.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl